♻️: refactor Gpt관련 상수값 Constant클래스에서 관리 #8
