+++ src/main/java/com/dongyang/core/domain/gpt/constant/GptConstant.java
@@ -2,6 +2,11 @@ package com.dongyang.core.domain.gpt.constant;
 
 public class GptConstant {
 
-	public static String MESSAGE_SYSTEM = "system";
-	public static String MESSAGE_USER = "user";
+	public final static String MESSAGE_SYSTEM = "system";
+	public final static String MESSAGE_USER = "user";
+	public final static String GPT_MODEL_NAME = "gpt-3.5-turbo";
+	public final static String MODEL = "model";
+	public final static String MAX_TOKENS = "max_tokens";
+	public final static String MESSAGES = "messages";
+	public final static String TEMPERATURE = "temperature";
 }
